requirements.txt
scipy
sklearn
pandas
pandas-datareader
matplotlib
pillow
tqdm
requests
h5py
pyyaml
tensorflow_hub
bayesian-optimization
spacy
gensim
flask
boto3
gym
keras-rl2
tensorflow-gpu
tables





